requirements.txt
# Utilities
requests
python-dotenv
markitdown
pyjwt


# AI 
openai

# Swarm
git+https://github.com/kevintupper/swarm_azure.git

# UI 
streamlit
  
# Security
msal